labeled unlabeled/mislabeled form fields on the Group Appointment Scheduler
This commit also adds an uniqid handlebars expression which generates an id
which is unique for each context in which it is called. This is done so that
that the template can be included multiple times, as long as it receives a
new context each time.

fixes CNVS-6608

Test plan:
- Confirm that all fields in the add or edit group appointment modal have
  labels that are read by screen readers.
- Confirm all labels read by screen readers are meaningful when read out of
  context, such as when tabbing through the form fields.
- Confirm that none of the changes made negatively impacted the dialog's visual
  appearance.
  - Confirm that the delete link also has a meaningful label.

Change-Id: Ia5b1d33e6d39266e44b8e369097128eaf69c0388
Reviewed-on: https://gerrit.instructure.com/26170
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Matt Fairbourn <mfairbourn@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>
Product-Review: Aaron Cannon <acannon@instructure.com>