boot fix for rails3 and compass
refs CNVS-9166

Sass::Plugin isn't automatically available by this point in the rails3
boot. but to my knowledge, it's late enough that it *can* be available,
so just require it.

rails2 has already loaded it by then, so this is a no-op there.

Change-Id: I14b6c160a4f35a9e711c4158418aa8a48e7a4c1f
Reviewed-on: https://gerrit.instructure.com/26232
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Product-Review: Jacob Fugal <jacob@instructure.com>
QA-Review: Jacob Fugal <jacob@instructure.com>