sync quizzes content_tag when publishing/unpublishing
When you publish or unpublish a quiz in a module it should also sync
the workflow state for its content_tag in that module so that the
sequence api doesn't think quizzes are published and show them to
students for parts of the code that use the sequence api ie: the
modules sequence footer.

fixes  CNVS-9118

Test Plan
With Draft State Enabled
1. Create three modules. Module A, Module B andModule C. Put a quiz in
each module.
2. In module A, unpublish the quiz in that module.
3. Navigate to module B and go to the quiz in the module. At the bottom
of the quiz there should now be a module sequence footer. That
footer should have a next button but not a previous button since the
quiz in the previous module isn't published.

Test Plan 2
1. Create a module.
2. Add a quiz to that module
3. Publish that quiz
4. Unpublish that quiz
5. Log in as a student and make sure you can see the quiz in the module
as a student but you can see it as a teacher.

Change-Id: Ibb4be0492aae33be9202f95e64cfbc3e09dc86b4
Reviewed-on: https://gerrit.instructure.com/26697
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Matt Fairbourn <mfairbourn@instructure.com>
Product-Review: Bracken Mosbacker <bracken@instructure.com>