remove quiz_submission loading from gradebook1
it's expensive and the only reason we need it is for the quiz id, so we can
link to the quiz submission page.  instead, include the quiz_id in the
assignment object and look it up from there.

fixes CNVS-9943

test plan:
- full regression test of gradebook1, specifically:
- make sure it works for regular assignments and quiz assignments
- make sure it works for students with quiz_submissions and without
- make sure grading students with and without quiz_submissions works
- make sure it works for courses where num students + assignments
  + submissions is < 100 and > 100

Change-Id: I84f4f543d8a021d005f8247c6863c23d794fb621
Reviewed-on: https://gerrit.instructure.com/27300
Reviewed-by: Cameron Matheson <cameron@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Caleb Guanzon <cguanzon@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>