speedgrader: always show group submission files
closes CNVS-9715

Some API consumers are uploading files intended for a group assignment
to a user's files endpoint instead of the group one.  This commit
allows a user's files to be associated with all submissions for an
assignment belonging to the same group.

NOTE: this does not allow users in the same group to steal each other's
files, only the owner of the file can submit it.

Test plan:
  1. create a group assignment
  2. upload a file to your user files area
  3. make a submission to the assignment with that file
  4. other users in that group should be able to view the file on their
     submission page
  5. try to submit the file from step 2 as another user in the group.
     this should fail

Change-Id: Ie949a5b71795fe0537305c57be94d48535605460
Reviewed-on: https://gerrit.instructure.com/27808
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Amber Taniuchi <amber@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>
Product-Review: Cameron Matheson <cameron@instructure.com>