fix for grade_change grader being nil or auto graded
The fix changes the grade change auditing so it will not create an index record
for autograded or nil grader events.

Note: Added a check for the event stream failure to raise the exceptions in the
test environment.  This can be overridden using the class as well.  This will
allow the failure spec to run as expected.

Test Plan:
  - Create a quiz.
  - Autograde the quiz for a student.
  - Check the grade change logs and they should exist.
  - No errors should be in the log table for the event stream

Change-Id: I682ab483950e123dc5522649df6ad43d6ff5a553
Reviewed-on: https://gerrit.instructure.com/27977
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jacob Fugal <jacob@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Clare Strong <clare@instructure.com>