allow accessible keyboard navigation in wiki-sidebar files tab
When inserting files many pages use wiki-sidebar. The files tab is now
accessible and allows navigation with the arrow keys and pressing enter.
There is also a dashed line that displays what area is currently being
focused on.

fixes CNVS-8997

Test Plan
With and without draftstate
Also, make sure you have files and folders created that are nested to
test this well.

1. Go to places in canvas that have wiki sidebars (editing a page for
instance).
2. Use keyboard arrows to navigate to the files tab on the wiki sidebar.
3. Use the tab key when on the Files Tab to start navigating the file
system.
4. Pressing the right arrow should expand a folder. Pressing it again
should go to its first child in the filesystem.
5. Pressing the left arrow should collapse a folder. If you press the
left arrow and are not on a folder it should take you to its closest
parent folder that is expanded.
6. Pressing up and down should navigate you to the next and previous
files/folders. When at the last file it will jump you to the next file
or folder in the file system.
7. Pressing enter on a file will insert the file into the text field
area. Focus will be in the text area.
8. There should be a dotted lined box that follows the focus on the
current file or folder.
9. It should also work in IE/Safari/Etc.Focus always remains on the
filesystem tree, it DOES NOT follow the currently selected items, we use
aria-selected tags to maintain this state for accessiblity. This is to
ensure Safari works sense it doesn't work when focus follows the
selected item.
10. When navigating with the keyboard scrolling the file section should
keep the highlighted or currently selected item in place.
11. Pressing the "end" button should take you to the last visible item
in the tree.
12. Pressing the "home" button should take you to the first visible item
in the tree.
13. When on the screen reader, folders should always be read with the
world 'folder' after their name and files will be read with 'file' after
their name.

Change-Id: I19bd5999298e96a21436f7e769e49bf3dbce9b29
Reviewed-on: https://gerrit.instructure.com/27616
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Nathan Rogowski <nathan@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Product-Review: Bracken Mosbacker <bracken@instructure.com>