fix append_base_url for some versions of nokogiri
The spec testing this code was failing in travis-ci, and I tracked it
down to a bug with the selector -- it's not quite css, but not quite
xpath either. It works or not, depending on the underlying version of
libxml that Nokogiri is using. I was able to reproduce the failure using
a recent 64-bit ubuntu VM.

Change-Id: I18838e6132ef495c909329b0a33fb7bb439b2b5d
Reviewed-on: https://gerrit.instructure.com/29181
Reviewed-by: Zach Pendleton <zachp@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Brian Palmer <brianp@instructure.com>
Product-Review: Brian Palmer <brianp@instructure.com>