add course settings to update and created activity logs
refs: CNVS-392 (g/27807)

Since the settings for a course is stored in a hash it was not triggering
that changes were made on create or update.  This fixes it so that course
settings changes are also tracked in the course activity logs.  This also
fixes an issue where indexed was always showing as changed due to its
default value.  This fixes it so it will only set it when required.  Same
as locale.

Test Plan:
  - Change some course settings during create and update.
      hide_final_grade
      hide_distribution_graphs
      allow_student_discussion_topics
      allow_student_discussion_editing
      lock_all_announcements
      large_roster
      public_syllabus

  - Make sure those are included in the course activity log.

Change-Id: I798cfc255d683339ab69a2daba14495c5f15d7e4
Reviewed-on: https://gerrit.instructure.com/29394
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jacob Fugal <jacob@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Nick Cloward <ncloward@instructure.com>