enable screenreader access to the publish/unpublish buttons in modules
When you go to the modules page and you have a module that has points
and a due date, the publish/unpublish cloud was not accessable to
screenreaders. It would hang up on the course pts and not read anything
else. This fixes that so you get the points read as well as are able to
get to the publish/unpublish cloud.

fixes CNVS-11514

Test Plan:
- Visit a modules page with a screen reader and draft state enabled.
- Add a module with points and a due date.
- Verify that the screen reader can read and monipulate the
  publish/unpublish button.
- Do the same on the assignment, quiz and discussion index pages.
- Make sure the screen reader reads the points possible and due date.

Change-Id: Iadf96a113fa5f708915b0861dadcd0f77cd1d47a
Reviewed-on: https://gerrit.instructure.com/31451
Reviewed-by: Bracken Mosbacker <bracken@instructure.com>
Product-Review: Bracken Mosbacker <bracken@instructure.com>
QA-Review: Matt Fairbourn <mfairbourn@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>