fix namespace issue with attachment relationship on quiz submissions
fixes CNVS-11986, CNVS-11672

test plan:

the problem on this one comes up when we are trying to view an attachment that
was created before the namespace change. So we have to massage the data in the
console to reproduce what's going on here, or test it with data that was created
before the namespace changes.

- as a teacher
  - create a quiz with a file upload question type

- as a student
  - take the quiz and upload the file

- in the console
  Attachment.update_all("context_type='QuizSubmission'", "context_type='Quizzes::QuizSubmission'")

- as a teacher
  - visit the moderate page for this student's quiz, and the attachment link
    should be visible to download the file.

Change-Id: I329bb422770e0d22fa33be189c0229298c12f759
Reviewed-on: https://gerrit.instructure.com/32319
Reviewed-by: Josh Simpson <jsimpson@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Caleb Guanzon <cguanzon@instructure.com>
Product-Review: Clare Strong <clare@instructure.com>