Upgrade MediaElementJS to 2.14.12
This commit just updates the vendor mediaElementJS code
so that we can see clearly what changed from upgrading
and what our changes are (yes, sadly we have to make
modifications to mediaelement-and-player.js's source)

so don't focus on QAing this in isolation. the actual changes
come in the commit after this.

refs: #CNVS-12293

Change-Id: I71f18f8dc8f30acd5cbb0730b1a97065e2aed2e3
Reviewed-on: https://gerrit.instructure.com/34015
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Clare Strong <clare@instructure.com>
Product-Review: Clare Strong <clare@instructure.com>
Reviewed-by: Paul Hinze <paulh@instructure.com>