fix multiple due dates from overlaping in modules
fixes CNVS-12637

When you go to the modulese page if you had multiple due dates it would
overflow into the pts column. This no longer does this.

Test Plan
---------
As a teacher
Given I have multiple sections set up in a course
And I've added an assignment to the modules page
And that assignment has multiple due dates on it
And that assignment has points on it
When I go to the modules page
I should the module item with multiple due dates text
And it should not be overlaping the points column

Change-Id: I77557f10eb83a1acca4aa21ad4b9d7ab1de2004d
Reviewed-on: https://gerrit.instructure.com/34387
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Clare Strong <clare@instructure.com>
Product-Review: Hilary Scharton <hilary@instructure.com>